Add security to management of processes through supervisord
As part of this commit we do following:
a) Allow supervisorctl to connect to supervisord only through unix
domain sockets. We also make the permissions of supervisord unix
domain sockets restrictive so that only root can invoke supervisorctl
command.
b) Make changes in nodemgr so that we do not depend on Juniper
proprietary extensions to supervisord.

Change-Id: Id0f351af0e1d5d1bf944cd70eaf0e649dddb2e56
Closes-PR: 1383939
Closes-PR: 1383503