Closes-Bug: #1566558 - SM support for External Openstack Provision
This check-in allows Server Manager to support External Openstack.
The SM can provision a node with only Openstack role and also a node with all other nodes.
These nodes can then communicate with the external openstack if they are configured correctly.

The config for both Openstack node and Cfgm node is given in the bug id.

Closes-Bug: #1566558 : New params name change caused error in External Openstack Support

The name change meant the validation was broken
Related-Bug: #1588999

Patch2: Typo in params, ip setting

Change-Id: I27a6c3cae121281f6996da38cef062c18dd3fcef