Closes-Bug: #1566558 - SM support for External Openstack Provision
This check-in allows Server Manager to support External Openstack.
The SM can provision a node with only Openstack role and also a node with all other nodes.
These nodes can then communicate with the external openstack if they are configured correctly.

The config for both Openstack node and Cfgm node is given in the bug id.

Change-Id: I27a6c3cae121281f6996da38cef062c18dd3fcef