Closes-Bug: #1635978 - Ceilometer param changed in puppet but not SM side
The parameter openstack::ceilometer::mongo::password: was changed to openstack::ceilometer::mongo:

But it is still being generated as openstack::ceilometer::mongo::password: in SM Lite
The hieradata template file was changed to reflect this.
This problem will be seen in SM lite or in SM if this param is not specified.

In SM, the cluster.json being used should use ceilometer:mongo: as here:

                    "openstack": {
                        "ceilometer": {
                            "mongo": "<password>",
If it is being set at all.

Change-Id: Iac6524b81665e53b5356a9b9ffe9fb5c6a436b13