SM-Provision-Sequence: write all flags for a host in single shot.
Closes-Bug: #1591267

Gather all tuples from a steps for a host and write them to the yaml file in
single shot.

Previously, SM was writing yaml files multiple times for same hostname. This
causes post_provision and compute to enable first and storage was in last of
the queue, that will written to yaml at a later point of time. on a over
loaded system, this could be 6-7 seconds also. In between, puppet could read
the yaml and get half-baked files.

Conflicts:
	src/server_mgr_main.py
Change-Id: I676354d74fe35252b82828ec214e66216058441f