Closes-Bug: #1641850 - Do DHCP Subnet validation only if feature is used
If feature is not used, the manually edited dhcp template file is taken as correct.
Also, the dhcp list for validations was not being correctly populated.
Once the user starts using the dhcp management feature we have to continue using that and cannot go back to manual file without deleting all hosts and subnets and then pasting the manually editing files and restarting.

Change-Id: I6173bc06478174959846e9a822da1588d5a0a3dc
Related-Bug: #1634275 - DHCP Feature