Partial-Bug: #1676560 - Testbed support for kernel_upgrade flag
This boolean flag is set under contrail_4 dict in server params if specified in testbed.py (for all servers)

In testbed.py please specify:

env.kernel_upgrade = True
env.kernel_version = "3.13.0-106"

Kernel_version is NOT mandatory is kernel_upgrade is given.
It will pick the default target kernel_version and upgrade/downgrade to that.

For trusty: 3.13.0-106
For xenial: 4.4.0-38

In server JSON it looks like:

{
    "cluster" : [
        {
            "parameters" : {
                "provision": {
                    "contrail_4": {
                        "kernel_upgrade": true,
                        "kernel_version": "3.13.0-106"
                    }
                }
            }
        }
    ]
}

Default value is False

Change-Id: I9a7231e7f6844af14c79eb91f4fe37a382b08cc0