Ability to add images using same tarball but with different image ID
Change the algorithm used to derive the image-id of the newly loaded container.
Previously, it was arrived at by creating a "diff" between the "docker images"
output before and after loading a container. This was preventing us pushing
images with the same docker image-id into the regsitry.
Now we adopt the more deterministic method of using the metadata.json
file inside the container tarball to identify the image id for that container.

Change-Id: Ib5e718371165c6ec2f43c2bba4beda8c36c632cd
Closes-bug: #1703459